+++ CHANGELOG.md
@@ -45,4 +45,5 @@
 
 1.0.9.2
 =============
+- FAQ: add model and helper
 - fix mage/cookies